% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jaccard.R
\name{Jaccard}
\alias{Jaccard}
\title{Jaccard}
\usage{
Jaccard(Data)
}
\arguments{
\item{Data}{A data.frame with 2 columns. The first refers to the set and the second the elements}
}
\value{
a data.frame with the set names and their Jaccard index
}
\description{
Calculates the Jaccard index between different sets.
}
\examples{
set.seed(123)
Data = data.frame(Class = sample(c("X", "Y", "Z"), replace = TRUE, size = 50),
                  Element = sample(LETTERS[1:15], replace = TRUE, size = 50))
Data = unique(Data)
Jaccard(Data)
}
