% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity_avr.R
\name{proximity_average}
\alias{proximity_average}
\title{Proximity from target to source}
\usage{
proximity_average(G, source, targets)
}
\arguments{
\item{G}{The original graph (often an interactome).}

\item{source}{nodes from the network (in a drug repurpusing set-up those are the disease genes)}

\item{targets}{targets in the network (in a drug repurpusing set-up those are the drug-targets)}
}
\value{
the proximity value for the source-targets
}
\description{
Calculates the proximity (average or closest) from source to targets.
}
\examples{
#' set.seed(666)
net  = data.frame(
Node.1 = sample(LETTERS[1:15], 15, replace = TRUE),
Node.2 = sample(LETTERS[1:10], 15, replace = TRUE))
net$value = 1
net =  CoDiNA::OrderNames(net)
net = unique(net)

g <- igraph::graph_from_data_frame(net, directed = FALSE )
T = c("G", "A", "D")
S = c("C", "M")
proximity_average(g, source = S, targets = T)
}
