% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Separation.R
\name{separation}
\alias{separation}
\title{Separation}
\usage{
separation(G, ST)
}
\arguments{
\item{G}{The original graph (often an interactome).}

\item{ST}{Set-Target data. It is a data.frame with two columns. ID and Target.}
}
\value{
the separation and distance of modules.
}
\description{
Calculates the separation of two set of targets on a network.
Often used to measure separation of disease modules in a interactome.
Separation is calculated as in Menche, J. et al (2015) <doi:10.1126/science.1257601>.
}
